+++ BreweryMaster/BreweryMaster.API/Recipe/Models/RecipeDetailsResponse.cs
@@ -6,10 +6,10 @@ namespace BreweryMaster.API.Recipe.Models
     {
         public int Id { get; set; }
         public required string Name { get; set; }
-        public float? BLGScale { get; set; }
+        public decimal? BLGScale { get; set; }
         public int? IBUScale { get; set; }
-        public float? ABVScale { get; set; }
-        public float? SRMScale { get; set; }
+        public decimal? ABVScale { get; set; }
+        public decimal? SRMScale { get; set; }
         public int? TypeId { get; set; }
         public string? TypeName { get; set; }
         public int? StyleId { get; set; }
@@ -19,13 +19,13 @@ namespace BreweryMaster.API.Recipe.Models
         public int? EvaporationRate { get; set; }
         public int WortVolume { get; set; }
         public int? BoilLoss { get; set; }
-        public float? PreBoilGravity { get; set; }
+        public decimal? PreBoilGravity { get; set; }
         public int? FermentationLoss { get; set; }
         public int? DryHopLoss { get; set; }
         public int? MashEfficiency { get; set; }
-        public float? WaterToGrainRatio { get; set; }
-        public float? MashWaterVolume { get; set; }
-        public float? TotalMashVolume { get; set; }
+        public decimal? WaterToGrainRatio { get; set; }
+        public decimal? MashWaterVolume { get; set; }
+        public decimal? TotalMashVolume { get; set; }
         public IEnumerable<RecipeFermentingIngredientResponse>? FermentingIngredients { get; set; }
     }
 }
